[utils] Fix `--match-filter` float comparison (e.g. `aspect_ratio`)
The `_match_one` function was using `int()` to parse numeric
comparison values, which caused float values like `1.78` to fail
parsing and fall through to `parse_filesize`, incorrectly rounding
them to the nearest integer.

This changes `int()` to `float()`, matching the behavior already
used in `_build_format_filter`. Fixes `--match-filter aspect_ratio=1.78`
and similar float-based filters.

Closes #16329

+++ yt_dlp/utils/_utils.py
@@ -3271,7 +3271,7 @@ def _match_one(filter_part, dct, incomplete):
             # and process comparison value as a string (see
             # https://github.com/ytdl-org/youtube-dl/issues/11082)
             try:
-                numeric_comparison = int(comparison_value)
+                numeric_comparison = float(comparison_value)
             except ValueError:
                 numeric_comparison = parse_filesize(comparison_value)
                 if numeric_comparison is None:
